requirements.txt
streamlit
google-generativeai
faiss-cpu
numpy
python-dotenv
pandas
plotly
langchain-text-splitters
langchain-core
